First commit of a prototype tool to parse DTrace provider definitions to Rust
- Adds dtrace.pest, a file for pest.rs defining the basic grammar of a
DTrace provider. Only basic arguments are supported, no pragmas, no
actions or predicates, only focusing on the provider definition.
- Adds the parser and tests of it in dtrace.rs
- Adds a function-like procedural macro `dtrace_provider!()`, which
parses a single .d provider definition file and generates a dummy Rust
struct. It's a unit struct, but with an impl block that defines a public
function that matches the DTrace probe in name and signature. The body
is currently empty.
- Adds a README