Make me a workspace
This is mostly NFC, just reorganizing this into a workspace. Proc-macro
crates are a bit weird, and can only export the exact proc macro
function they define. The convention here is to have two closely related
crates, one of which defines/exports the macro, and a public one that
just re-exports the macro along with any other related Rust code as a
normal lib crate.

Note that this also moves the simple.rs example into the expected place
so we can `cargo run --example simple`. It also does a slightly better
job handling errors when the provider.d file can't be found, explaining
to the user that the path has to be relative to the project root.