Adds static library implementing the actual DTrace probe call
- Adds the usdt-macro/native subdirectory, which contains a simple
provider definition, C source and header, and a Makefile. These are
built into a static library as one would normally do with DTrace. The
actual proc-macro now contains an extern declaration of this probe
function, and specifies how it should be linked to. The Rust-side of the
probe function now performs an FFI call to the C function, and thus the
DTrace probe.
- Updates README with new build details. This is currently pretty
annoying, since using the library in any meaningful way requires telling
cargo how to find the static library. Hopefully we can streamline this
in the future.