Factors out the DTrace parser into its own crate.
This commit moves the code for parsing DTrace providers into its own
crate within the workspace. It may prove useful in the future to have an
API specifically for working with providers, probes, etc, rather than
directly working with the tokens exposed by the grammar. This is a first
step in that direction. The next would be to construct one or more types
that explicitly represent the provider, probes, and arguments of the
probe. These can be pulled from the parsed tokens internally, and then
exposed for quoting and converting to Rust tokens in the macro. This
will make the macro much simpler, and allow us to use the
provider-parser for other purposes (e.g., generating a C implementation
from it, in addition to the Rust impl).