Adds a more full-featured D parsing library crate
- Adds Rust types representing the various components of a D provider
file. This includes the data types, probes, providers, and files.
- Adds methods for emitting C and Rust type names for each probe
argument, and emitting the Rust code that _converts_ from a Rust to C
type (used in the FFI calls).
- Adds methods for emitting full Rust implementations, C declarations,
and C definitions for probes/providers/files. These are used to generate
the various bits of glue code needed to make a static library that calls
DTrace probes, the Rust FFI declarations of those functions in the
library, and the Rust implementations/wrappers that actually call those
FFI functions.
- Adds a binary to the usdt crate, which can be used to dump the result
of parsing a DTrace provider file, in either Rust or C
declaration/definition format.
- Adds a subcommand to the above binary, which can be used to dump out a
build.rs script. The artifacts built by this script are the C glue code
which the Rust code calls into via FFI. This allows us to generate a
static library for _each_ provider, rather than a static library shared
by _every_ provider, with a minimum of fuss. Users have to install this
tool and place the build script in their code, which is kind of
annoying, the tradeoff we have to make for now.
- Rewrites the actual macro to use the dtrace-parser library
- Removes the implementation of the macro directly using the parsed
code, instead using the dtrace-provider crate.
- Removes the native library code from the macro crate
- Removes the simple example, and adds a standalone executable that
functions as an example in its own crate
- Updates README to describe the current process for defining, building,
and using probes in Rust. Includes a note about installing the
executable for easier use.