Moves the Rust side of the probe call to a macro
- The Rust generated code was previously a struct and some associated
functions, so the probe would be called with `provider::probe()`. This
has some nice properties, but the ultimate goal is probably to switch
the internals here to some inline assembly. The function call may be
inlined, especially since it does nothing but forward to the FFI
function, but there are no guarantees about that. This commit swaps that
to a macro to ensure that there is no call boundary to worry about. Note
that this has some unpleasant side effects, as macros are basically
always located at the crate root (i.e., you can't do
`provider::probe!`), and only _after_ they've been defined. This means
users must be sure to put the `dtrace_provider!` macro call textually
before any invocations of the probes. So it should really always be the
first thing, including before any submodules are defined in a crate.
- Updates the README
- Updates the probe-test to use the macro invocation.