Better string handling
- Previously the code unconditionally created a new CString from a
string type passed to the macros, and panicked if that was not possible.
This commit does better, but it's also sneaky. In place of an actual
char *, we pass a transmuted pointer to the Rust string type, which is
guaranteed to be a tuple of (data pointer, size). Inside the C function,
we copy `size` bytes from the data pointer into a newly-allocated
buffer, and null-terminate it. This guarantees that we only copy if
needed, since its behind the is-enabled conditional. It's also the
safest we can be in this situation: Rust strings may have intervening
null bytes, so the string may be _shorter_ than the user expects, but
it'll never be longer, since it is null-terminated. We also just assert
if the allocation fails.
- Puts calls to the actual probe function behind the is-enabled probe.
- Removes floating-point probe argument types.
- Expands the probe-test to showcase strings.
- Some miscellaneous formatting of the emitted code, and adds data files
with the expected, formatted output of the various code generation
routines.