Adds basic inline ASM version which sends DOF via ioctl
- Adds support for serializing/deserializing the argument mappings
section in a DOF
- Adds mac and non-mac ASM for pushing probe records into a special
object file section
- Adds mac and non-mac function for sending this data to the DTrace
in-kernel code via ioctl(2)
- Confirmed that this works on both mac and illumos. Note that arguments
are not yet supported. "Works" here means that, on illumos, the probes
are registered and fired correctly; and, on mac, the probes are
registered correctly, but not fired, presumably because Apple handles
the addresses and offsets in the linker and kernel in very odd ways.