Adds an Apple specific implementation, using the platform linker to do the heavy lifting
- Splits out the `usdt-impl/src/asm.rs` module into a directory, with
`mac.rs` and `other.rs` files that are conditionally included based on
the platform.
- The actual implementation on mac is quite different. The process of
compiling the probes first calls out to `dtrace(1)` to get the
apporpriate C header file from the provider definitions. This file is
parsed to collect the names of the relevant symbols for each provider
and probe. These look like
`__dtrace_probe$provider_name$probe_name$xxx`, where the last bit is a
hashed representation of the probe's argument types. The probe,
is-enabled probe, and stability/typedef symbols are all collected, and
inserted into the generated macro. This is somewhere between the pure
ASM record version implemented for other platforms, and the original FFI
implementation: we're actually declaring an extern "C" FFI function and
calling that, but Apple's linker is manipulating this at link and load
time to do the work the other platform implementation does manually.
- Adds basic tests of the internal string-munging methods in the macos
implementation.