WIP: Adding cross-platform inspection tool
- Removes the `build.rs` and bindgen depedency in the `dof` crate. The
actual DOF types are very unlikely to change, as they're a long-standing
on-disk format. Fewer dependencies, faster builds, and more supported
platforms are a worth the tradeoff here.
- Removes the binary from the `dof` crate, and moves the functions for
extracting pretty-printed DOF data into the `dof::fmt` module. It is
public, and used in the inspection tools elsewhere.
- Moves the functionality for processing the custom ASM records into a
separate, cross-platform module. This is used in `no-linker.rs` to
extract the records and send to the DTrace kernel module, and also in
the main `usdt` binary for inspecting the probe records.
- Adds a binary target to the main `usdt` crate, which dumps out
information about either DOF sections (e.g., from a file built on
macOS), or the ASM records this crate inserts (on other platforms). Note
that this may not accurately resolve the probe's function name when
parsing ASM records, as `dladdr(3)` queries the platform's dynamic
linker.