Adds support for arguments in ASM records on no-linker platforms
- Adds an argument count and strings for each argument to the ASM
records built in `usdt-impl/src/no-linker.rs`. This is compatible with
the previous version 1, so it's not currently bumped here.
- Adds support for deserializing these arguments from the ASM record and
inserting them into the `dof::Probe` objects built from the records.
- Small bugfix in DOF serialization method
- Adds method to convert a `dtrace_parser::DataType` into a C type
string, used in building ASM records.
- This has been tested on mac and illumos. In both cases, the arguments
are correctly represented with `dtrace -lv`; the actual probe-test
examples run correctly; and the arguments are correctly printed with
`dtrace` when tracing the arguments. The `usdt` binary can correctly run
on mac, illumos, and linux, with the obvious caveat that the function
address will not be resolved to a name if the object file was not built
for the current platform.