Adds no-linker feature for building on macos without relying on the platform dynamic linker
- Adds the `"no-linker"` feature flag to the `usdt` and `usdt-impl`
crates, forwarding the former to the latter
- Better conditional compilation of the `usdt_impl::internal` module,
based on the platform and new set of feature flags
- Combines some of the shared code used to generate the USDT macros
themselves into a `usdt-impl/src/common.rs` mod. The actual
implementation of the macro in the macos/linker case has been updated.
Previously, we were directly calling the (linker-defined) probe function
by creating a Rust fn with link-name of that mangled probe function.
This mapped directly onto the "usual" C implementation. That has been
updated to marshal the arguments just as we do on no-linker platforms,
and use the x86 call instruction directly. This all works, but it is a
difference, with the goal of being able to share code and test
everything more directly.
- Implements the linker/no-linker versions using the new shared code
- Adds tests for the new shared code
- Adds support for correctly extracting the module name in the no-linker
version of the code. This works by returning the containing file name in
addition to the function name from the `addr_to_info` function, and then
passing that as the module name in the DOF sent via ioctl(2).
- Adds simple test for the `asm_rec` function