Adds tests to verify useful error messages when compiling fails (#13)
- Adds a suite of tests which verify the compiler error messages when
the `usdt` crate is misused in various likely ways. This includes a
provider definition with an unsupported type, a type mismatch in the
probe macro invocation, calling the probe macro without a closure, and
specifying a non-existent provider file to the proc macro or build.rs.
- Adds back the `asm` feature as a default and updates the CI config
accordingly.
- More appropriate conditional compilation of the actual internal
implementation file under various combinations of platform a feature
flags.