this commit is broken in an interesting way and also in a tedious way:
If you cd to tests/fake-cmd and cargo run you'll see different results
depending on whether you do or do not specify --release. Without the
flag our ELF section is included in the final binary; with the flag
the section is there but is zero-lengthed.

The tedious way in which this is broken is that it attempts to force
the creation of our ELF section like this:
```
    #[link_section = "set_dtrace_probes"]
    #[used]
    static FORCE_LOAD: [u64; 0] = [];
```
This sort of works, but fails if a probe is used more than once.