Prunes the dependency tree (#15)
- Removes extraneous structopt dep from `dof` crate
- Adds a "des" feature to many of the packages. This is used to gate
inclusion of the tools for deserializing the probe records or DOF itself
from an object file. The effect on the dependency tree is to make the
large `goblin` dependency optional.
- Adds a small, separate binary crate `dusty`, which has the binary for
actually reading the DOF or ASM records from an object file. This is
really just the former `usdt/src/main.rs`, with appropriate changes to
the feature flags and depdendency tree.
- Note that the `"dof/des"` flag is not a default feature. It is opted
into via the `"usdt/des"` feature, which is _not_ implied by the "asm"
and "no-linker" features of the same crate. This allows the
deserialization tooling to be opted _out_ of, e.g., by most users of the
`usdt` crate. Note that the usdt::record module is conditionally
included on the basis of platform and the no-linker or asm features.