Fixes bad dependencies related to the dof crate
We previously tried to make the `dof` crate optional, but this wasn't
quite right. It's optional if the `no-linker` feature is _not_ enabled
_and_ one is not on a platform where that implementation is selected
regardless of the feature. One obvious case is illumos.

Unfortunately, platform-specific features are not yet implemented. This
commit instead makes the `dof` crate an optional dependency on macOS
platforms, and a required dependency on all other platforms. Note that
the `dof` crate is still pulled in on macOS if the `no-linker` feature
is explicitly enabled.

This commit also removes the `--features no-linker` bits from the
buildomat script, and bumps versions in preparation for a release with
this new dependency graph.