Fixes type checking for probes with zero arguments
The internal type-checking mechanisms were broken for probes with no
arguments. One could write an argument closure returning _any_
arguments, and this would "pass" our internal type-checking, as we only
unpacked the _expected_ arguments from the closure. As there were no
expected arguments, there was no type-checking.

This commit modifies type-checking to verify that if the probe expects
zero arguments, we can assign the result of calling the argument closure
to a variable of type `()`. It also modifies the type checking of
serializable types. Before we used a generic parameter with a trait
bound `T: ::serde::Serialize`. This is actually not enough, since it
would successfully type-check an argument closure that returns a
_different_ type that is serializable than the one in the probe
signature. We now collect the actual, concrete type from the probe
signature during parsing, and insert that into the signature of the
type-check function as we would any other concrete type.

Two new try-build tests have been added that verify this behavior.