Improves type-resolution in provider modules.
Type resolution in providers defined as Rust modules was subtly broken.
In particular, one needs to import types, such as `super::Foo`, into the
module for them to be referenced in probe functions. However, we cannot
use that exact use statement, or the imported type (`Foo`) in the
generated probe macros -- in general it won't be available at all, let
alone at the same path, and relative paths are meaningless as the macros
may be invoked from anywhere in the crate.

This commit adds better support for resolving these types, and
re-importing them as necessary into the scope of the generated probe
macros.

The Provider type has been augmented to carry around the tokens from any
`use` statements in a Rust provider module. These statements are then
re-emitted verbatim into the scope of the probe macros, so that they can
be referred to in the same way (or any other valid way at that scope).
This explicitly disallows using `super` in these module-level imports,
and a test has been introduced to verify this.

The original, verbatim module that consumers write is also now
re-emitted. Specifically, we previously chucked that in the bin,
emitting only the macros and type-checking functions. We now add those
latter two _into_ the original provider module. This helps with
resolution or things like Rust analyzer, trying to follow the macro
rule-of-thumb to emit the original thing unchanged as much as possible
to support more helpful error messages in the event that the _original_
tokens are malformed in some way.