Separates dtrace-parser and main impl crate types
This commit separates out the `dtrace-parser` types `DataType`, `Probe`
and `Provider`, and those used in the `usdt-impl` crate. The former
doesn't need any notion of serializable types, while the latter does.
The former also need not depend on `syn`, while the latter does.

This also cleans up and improves the type-resolution even further. In
particular, the previous version would fail to resolve types in the
check-functions for serializability, if those types were renamed when
importing into the provider module. The code used to alwasy refer to
these as `use super::ImportedTypeName`, which doesn't work if the type
has been renamed. In fact, the use statement isn't needed, since these
check functions share a scope with the module itself. Any type names
that resolve in one place also resolve in the other.