Add default features to usdt-impl crate, fix name-length max
This fixes an annoying issue related to feature-detection. Specifically,
it seems that we need the `usdt-impl` crate to specify the `asm` feature
as a default for the `default-features = false` bit in the top-level
`usdt` crate to work. Without this, it's not clear why, but the
`usdt-impl/asm` feature may or may not be selected dependending on how
consumers of the `usdt` crate set their feature flags. Note that this
also means _any_ crate that names `usdt-impl` as a dep, including the
macro crates, must specify `default-features = false`.

This also fixes a bug where long names sent to DTrace result in an
EINVAL. Names like the provider are limited in length, usually to 64 or
128 characters (including the null-terminator). This ensures that
pulling out the probe records, we truncate the strings we send over to
DTrace via ioctl(2) are no longer than the maximum.

This also bumps the version numbers, in preparation for a release.