Misc fixes
Relaxes the version bounds in most Cargo.toml files to simplify
updating third-party dependencies and inter-workspace dependencies.

Adds an internal method for converting to JSON, and uses that in the
generated/quoted code. This is so users don't need to depend on
serde_json in their own Cargo.toml files, just usdt.

Removes some outdated platform-detection stuff in the no-linker.rs
implementation. This is no longer selectable on macos, so some
conditional compilation for that target OS has been removed.

Fixes small issue with the maximum length of a provider name string. The
PID is automatically appended by DTrace, so this takes that into account
when limiting the name stored in ASM records and shipped to DTrace via
the `ioctl(2)` call.