Support receiving probe argument closure data by either value or reference
This modifies the way in which the probe argument closure data is
received. Previously, data was expected to be of exactly the type
specified in a probe function parameter. E.g., a `&[T]` must be
returned from the closure as an actual `&[T]`, not something that
coerces to that type, such as a `Vec<T>`.

This commit modifies the probe closure type-checking function, to accept
things that can be referenced as the expected type: `AsRef<str>` for
anything like a string, `Borrow<T>` for any `T` or `&T`, etc. This
allows returning data more flexibly, such as by reference or value. A
common example is slices -- people often write function parameters as
`&[T]`, and expect to pass a `&Vec<T>` there, which works because
`Vec<T>` implements `AsRef<[T]>`. This commit implements such flexible
argument checking.

A new test and example has been added, in `tests/argument-types`, which
shows a fairly extensive enumeration of the types supported in both
probe function signatures and in argument closures.