Supports returning captured closure arguments by value
It wasn't previuosly possible to return data by value from a probe
argument closure, at least where the type does not implement Copy. This
was due to the fact that the argument closure seems to be invoked twice:
once during argument type-checking and once during the actual
implementation of the probe. This is not really true. The first call
site is never actually run, it's just there to verify the correct types
are returned from the closure. However, the compiler sees the closure,
which is `FnOnce`, called twice. The argument captured by value, but
which doesn't implement Copy, can't be returned twice.

This commit adds support for returning such values, by cloning the
closure before passing to the type-checking function. This requires that
the USDT types also implement clone, but it's important to note that the
type will _never actually be cloned_. It just needs to not generate a
compiler error to call the `clone()` method.