Always borrow probe data when possible (#35)
* Support receiving probe argument closure data by either value or
reference

This modifies the way in which the probe argument closure data is
received. Previously, data was expected to be of exactly the type
specified in a probe function parameter. E.g., a `&[T]` must be
returned from the closure as an actual `&[T]`, not something that
coerces to that type, such as a `Vec<T>`.

This commit modifies the probe closure type-checking function, to accept
things that can be referenced as the expected type: `AsRef<str>` for
anything like a string, `Borrow<T>` for any `T` or `&T`, etc. This
allows returning data more flexibly, such as by reference or value. A
common example is slices -- people often write function parameters as
`&[T]`, and expect to pass a `&Vec<T>` there, which works because
`Vec<T>` implements `AsRef<[T]>`. This commit implements such flexible
argument checking.

A new test and example has been added, in `tests/argument-types`, which
shows a fairly extensive enumeration of the types supported in both
probe function signatures and in argument closures.

* Supports returning captured closure arguments by value

It wasn't previuosly possible to return data by value from a probe
argument closure, at least where the type does not implement Copy. This
was due to the fact that the argument closure seems to be invoked twice:
once during argument type-checking and once during the actual
implementation of the probe. This is not really true. The first call
site is never actually run, it's just there to verify the correct types
are returned from the closure. However, the compiler sees the closure,
which is `FnOnce`, called twice. The argument captured by value, but
which doesn't implement Copy, can't be returned twice.

This commit adds support for returning such values, by cloning the
closure before passing to the type-checking function. This requires that
the USDT types also implement clone, but it's important to note that the
type will _never actually be cloned_. It just needs to not generate a
compiler error to call the `clone()` method.

* Updates docs and adds Clone trait bound to usdt-attr-macro type-checking