Module-scoped probe macros (#38)
* Moves to scoped macros, within a real provider module.

This changes the implementation from generating a single, crate-wide
macro named by concatenating the provider and probe, to a scoped
version. The provider is now a real module, and the probes are macros,
scoped to that module. For example, the macro `my_provider_my_probe` is
now `my_provider::my_probe`.

The naming here is configurable. Users may specify zero or more provider
module names, and the code will generate a series of nested modules for
each. The probes will be at the "bottom" of this series of modules. For
example, users can put their probes at a path
`{provider}::a::b::probe_{probe}`. This allows them to scope and
organize the macros how they wish.

This overall is much better, as it's more intuitive to name the provider
and probe, and hopefully more obvious. It does have several interactions
with existing tests, but in general the system behaves more similarly to
how users might expect other Rust items to behave. As an example, with
the default naming scheme, the probe function item and the generated
probe macro are the same, expect for the trailing `!`: `provider::probe`
at the definition site and `provider::probe!` at the call site.

* Updates scoped probe macros for macos