Adds documentation about new `asm_sym` feature
This commit adds documentation and fixes tests for the new `asm_sym`
feature, which gates symbol operands in inline assembly. These operands
are integral to this crate's implementation on macOS. Since we're using
a nightly toolchain, this is a breaking change, and there's no way to
build `usdt` or downstream crates that compile both before and after the
addition of this feature. The documentation describes this clearly, and
provides the feature flags needed on either side of this divide.

The tests have also been updated with the feature flag, as they run on
the latest nightly.