Supports crates using the new resolver v2
The dependency resolver v2 does not merge features in the same way as
prior versions. V2 is also the default as of Rust edition 2021. This
could lead to incorrectly building multiple versions of the usdt crate,
including one erroneously selecting the empty implementation.

This adds feature flags for the `usdt-macro` and `usdt-attr-macro`
crates that ensure we're building with the correct set of flags
regardless of which resolver is used.