Allows renaming the provider, rather than the probe module
After the change to module-scoped macros, we kept the ability to rename
the probe module from the earlier implementation. This meant the
procedural macros might rename the module, which is a bit confusing.
This switches to support renaming the _provider_, but not the module.
The module can be named to whatever the user wishes, and the probes are
always available in that scope. If the provider name is not chosen in
the attribute macro, it defaults to the module name, but this can be
overridden.