Moves extern symbol definitions into the macro bodies on macOS (#48)
On macOS, we define extern functions with valid Rust identifiers for a
few key bits of information about the providers and probes. These
symbols are generated by the macOS linker, but we collect their names
from the output of `dtrace -h` during macro generation.

Previously these symbols were defined in a separate scope that the
actual probe declarative macro bodies. For example, the stability symbol
was defined in the provider module, so one "level" of scope above the
probe macro bodies. This means we need to be able to refer to that
symbol by path, which is difficult, since the probe modules can
generally be anywhere in the crate at this point.

This moves those extern symbols directly into the scope of the probe
macro body. They can just be referred to without any other path
components on their names, since they're "right next" to the place
they're used.