FreeBSD: support nightly and stable(link-dead-code)
Provide a macro version of register_probes!() enables a explicit reference to
the set_dtrace_probes section from the crate consumer, and prevent lld from gc
the section. This trick along with the unstable "used(linker)" feature enable
use on FreeBSD target with nightly toolchain without explicit "link-dead-code"

Since the "used_with_arg" is available only on nightly, a patch is applied on
usdt-impl/build.rs such that standard backend is enabled only for FreeBSD
targets that are either on nightly or have link-dead-code explicitly passed to
the linker.