Remove register_probes! macro, and use R flag for sections instead
Turns out we do not need all the hacks to get FreeBSD linker working.
GNU and llvm both support the "R" flag for sections, which prevent
them from gc.

This make the ugly register_probes! hack obsolete so it can be removed.
No one deserve to have it in their codebase.

https://lists.freebsd.org/archives/freebsd-hackers/2022-June/001191.html