Add both path and version dependencies
A previous commit removed the version specification for the path
dependencies in the repository. This all works fine, until the crates
are published. Crates on crates.io may not have path dependencies, but
there is no version to use.

The accepted solution appears to be to specify _both_ path and version
dependencies. The former is used when building from source, the latter
when publishing.