Update `usdt::probe_records` to work on macOS
After failing to get an xtask experiment @bnaecker was working on
I realized that when we extracted out `probe_records` from dusty
we missed a key piece of functionality that enabled it to work on osx.

As I was working on this I realized we really needed to rethink how
formatting itself was happening. Now there's only one formatting logic
branch _and_ probe_records should work upstream in the xtask.

The one thing to note is that `probe_records` now returns a
`Vec<Section>` where it previously only returned a `Section`.