Fix race in JSON tests
The tests for JSON output work by starting DTrace as a subprocess;
firing some probes in the Rust test code; and then grabbing the DTrace
output and verifying its properties. There was a race here: although we
waited to fire the probes until requested, we never actually waited for
DTrace to correctly start and indicate readiness. Test hangs ensue,
at least on macOS where `sudo` actually requires a password in most
cases.

This adds synchronization in the other direction, waiting for DTrace to
print a sentinel in its BEGIN clause, so that the tests don't start
until it has successfully spawned.

Fixes #155