Extend illumos yeet mechanism to FreeBSD, removing FreeBSD-specific workarounds
Turns out, FreeBSD's linker would discard the `set_dtrace_probes`
section unless we use the trick which works on Illumos: have another
section `yeet_dtrace_probe` which would reference `set_dtrace_probes`
and prevent the linker from garbage-collecting it.

Unfortunatelly, on FreeBSD this has to be added on top of setting 'awR'
as the section parameters of `set_dtrace_probes`. The 'R' (Retain) and
yeet work together to retain this section.

In particular, this solution allows us to remove the FreeBSD-specific
features that would either use `#[used(linker)]` macro on nightly or
require `RUSTFLAGS="-C link-dead-code"` flag on stable.