Add FreeBSD CI
FreeBSD pipeline uses a 14.1 VM (arbitrary choice) in which we manually
enable DTrace, install Rust and then run the actual tests.

Also fixed test-unique-id flakiness on FreeBSD: thread IDs can be large
enough that when shifted left by 32 bits (as in UniqueId::materialize)
and formatted with '%d' (signed), they become negative and fail to parse
as unsigned integers in Rust. Changed to '%u' (unsigned) formatting to
fix this.